Closes #234
